Temporary fix problem with float conversion.

+++ les/ext/coin/osi_clp_solver_interface.py
@@ -16,6 +16,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import numpy
 import itertools
 
 from les.ext.coin import _osi_clp_solver_interface
@@ -58,7 +59,7 @@ class OsiClpSolverInterface(_osi_clp_solver_interface.OsiClpSolverInterface,
         for i, v in itertools.izip(row.indices, row.data):
           r.insert(int(i), v.astype(float))
         # NOTE: fix coef because of C++ signature
-        self.add_row(r, "L", problem.get_rhs()[p].astype(float), 1.)
+        self.add_row(r, "L", float(problem.get_rhs()[p]), 1.)
     elif details.get("rhs", True):
       pass
       for i in xrange(len(problem.get_rhs())):
